# grammars: 1.5x faster inference w/ complex grammars (vector reserves / reuses)