See discussion in https://github.com/ggerganov/llama.cpp/issues/4218#issuecomment-2049604659

Here's simple code to repro 1.6x faster inference speedup (on Metal) w/ a nested repetition-heavy grammar from https://github.com/ggerganov/llama.cpp/pull/6555 (for JSON schema `{"items": {"type": "number"}, "maxItems": 100}`):

```bash
git clone https://github.com/ochafik/llama.cpp --branch grammar-speedup3 llama.cpp-grammar && \
    cd llama.cpp-grammar && \
    git pull && \
    mkdir -p models/7B

echo '
    decimal-part ::= [0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9])?)?)?)?)?)?)?)?)?)?)?)?)?)?)?
    integral-part ::= [0-9] | [1-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9] ([0-9])?)?)?)?)?)?)?)?)?)?)?)?)?)?)?
    number ::= ("-"? integral-part) ("." decimal-part)? ([eE] [-+]? integral-part)? space
    root ::= "[" space number "," space number "," space number "," space number "," space number "," space number "," space number "," space number "," space number "," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number ("," space number)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)?)? "]" space
    space ::= " "?
' > json_numbers.grammar

hyperfine \
    --warmup 1 --runs 5 \
    -L branch grammar-speedup3,master \
    --setup 'git checkout {branch} && make clean && make -j LLAMA_CURL=1 main' \
    './main \
        -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf \
        --grammar-file json_numbers.grammar \
        -p "List of 20 integers starting from 0" \
        --seed 12344'
```

<details>
<summary>Show results</summary>

```
Benchmark 1: ./main -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf --grammar-file json_numbers.grammar -p "List of 20 integers starting from 0" --seed 12344 (branch = grammar-speedup3)
  Time (mean ± σ):      8.405 s ±  0.234 s    [User: 6.806 s, System: 0.412 s]
  Range (min … max):    8.179 s …  8.750 s    5 runs
 
Benchmark 2: ./main -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf --grammar-file json_numbers.grammar -p "List of 20 integers starting from 0" --seed 12344 (branch = master)
  Time (mean ± σ):     13.386 s ±  0.109 s    [User: 9.596 s, System: 2.552 s]
  Range (min … max):   13.253 s … 13.520 s    5 runs
 
Summary
  ./main -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf --grammar-file json_numbers.grammar -p "List of 20 integers starting from 0" --seed 12344 (branch = grammar-speedup3) ran
    1.59 ± 0.05 times faster than ./main -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf --grammar-file json_numbers.grammar -p "List of 20 integers starting from 0" --seed 12344 (branch = master)
```

</details>

cc/ @HanClinto