@HanClinto 2024-04-11
Similar to the integration tests, `examples/gbnf-validator` will eventually need to be updated to use the new API as well. That's lower priority though, and I can do that after we get through all of this.

Reading through this PR, I'm amazed that such a simple change provides such a dramatic speedup. I still just have a hard time believing it's as effective as it is. :)