@ochafik 2024-04-11
> Similar to the integration tests, examples/gbnf-validator will eventually need to be updated to use the new API as well.

Done, thanks!

> Reading through this PR, I'm amazed that such a simple change provides such a dramatic speedup. I still just have a hard time believing it's as effective as it is. :)

Yeah I tried half a dozen similar rewrites and only these lucky two struck a chord :-D (let's hope for much more dramatic speedups w/ upcoming changes https://github.com/ggerganov/llama.cpp/issues/4218#issuecomment-2049604659)