@HanClinto 2024-04-11
FWIW, I've independently confirmed the (rather dramatic) speedup results of **1.71x** on my system (wow!):

```
Benchmark 1: ./main \
        -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf \
        --grammar-file json_numbers.grammar \
        -p "List of 20 integers starting from 0" \
        --seed 12344 (branch = grammar-speedup3)
  Time (mean ± σ):     12.302 s ±  1.341 s    [User: 8.369 s, System: 1.672 s]
  Range (min … max):   11.405 s … 14.642 s    5 runs

Benchmark 2: ./main \
        -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf \
        --grammar-file json_numbers.grammar \
        -p "List of 20 integers starting from 0" \
        --seed 12344 (branch = master)
  Time (mean ± σ):     20.978 s ±  1.003 s    [User: 11.519 s, System: 6.894 s]
  Range (min … max):   19.908 s … 22.488 s    5 runs

Summary
  ./main \
        -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf \
        --grammar-file json_numbers.grammar \
        -p "List of 20 integers starting from 0" \
        --seed 12344 (branch = grammar-speedup3) ran
    1.71 ± 0.20 times faster than ./main \
        -mu https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf \
        --grammar-file json_numbers.grammar \
        -p "List of 20 integers starting from 0" \
        --seed 12344 (branch = master)
```

Macbook Pro, Apple M1 Pro, 32 GB of RAM, and about a billion open Firefox tabs.

Really awesome work, @ochafik !